When reload is requested, ask the user to reload via an alert (#3104)
