+++ ui/main/src/app/modules/core/reload-required/reload-required.component.scss
@@ -1,0 +1,50 @@
+/* Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * See AUTHORS.txt
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ * This file is part of the OperatorFabric project.
+ */
+
+ .opfab-info-message {
+  text-align: center;
+  width: 100%;
+  position: fixed;
+  z-index: 4000;
+  left: 0;
+  top: 0;
+  color: white;
+}
+
+.opfab-alert-error {
+  background-color: #e87a08;
+  font-size: 16px;
+  font-weight: 600;
+  text-align: left;
+  height: 60px;
+  line-height: 60px;
+  padding-left: 30px;
+}
+
+.opfab-alert-close {
+  position: absolute;
+  top: 0;
+  right: 0;
+  z-index: 2;
+  color: var(--opfab-bgcolor);
+  margin-right: 30px;
+  font-size: 30px;
+  cursor: pointer;
+}
+
+.opfab-alert-close:hover,
+.opfab-alert-close:focus {
+  color: var(--opfab-navbar-color);
+}
+
+.opfab-reload-link {
+  text-decoration: underline; 
+  cursor: pointer;
+}
+
