+++ ui/main/src/assets/i18n/nl.json
@@ -74,7 +74,6 @@
     "activateSoundText":"Geluid is niet geactiveerd, klik om te activeren",
     "activateSoundButton":"Activeer geluid",
     "sessionExpiredText":"De sessie is verlopen",
-    "reloadRequestedText": "Applicatie herladen vereist",
     "disconnected": "Je bent losgekoppeld.",
     "disconnectedByNewUser": "De verbinding met u is verbroken omdat een andere gebruiker met dit account heeft ingelogd.",
     "disconnectByAppLoadedInAnotherTab": "De verbinding is verbroken omdat een ander tabblad van deze browser werd gebruikt om in te loggen met deze URL"
@@ -480,5 +479,9 @@
     "december": "Dec",
     "week": "Week",
     "placeholder": "JJJJ-MM-DD"
+  },
+  "reloadRequired" : {
+    "message": "Applicatie herladen vereist. Om te herladen klik ",
+    "linkText": "hier"
   }
 }
